moving legacy level-control cluster impl into codegen/
