+++ src/app/clusters/level-control/app_config_dependent_sources.gni
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "level-control.cpp",
-  "level-control.h",
+  "codegen/level-control.cpp",
+  "codegen/level-control.h",
 ]
